ui前

+++ pigeon_cloud_breed/src/main/java/com/dove/breed/entity/ShipmentEntryBase.java
@@ -46,7 +46,7 @@ public class ShipmentEntryBase extends Model<ShipmentEntryBase> {
     //产品编号
     @ApiModelProperty(value = "产品编号")
     @TableField("type_id")
-    private Long typeId;
+    private Integer typeId;
 
     //名称
     @ApiModelProperty(value = "名称")
